Revert "Removed escapeHtml method from helpers/html"
This reverts commit aaadba62c066a32743725c0ca852db68a342fdcb.
