+++ src/post/content.unit.test.ts
@@ -44,36 +44,13 @@ describe('ContentPreparer', () => {
 
         describe('Escaping HTML', () => {
             it('should escape HTML', () => {
-                const testCases = [
-                    {
-                        input: '<p>Hello, world!</p>',
-                        output: '&lt;p&gt;Hello, world!&lt;&#x2F;p&gt;',
-                    },
-                    {
-                        input: 'Lorem ipsum dolor <img src="https://example.com/image.jpg" />',
-                        output: 'Lorem ipsum dolor &lt;img src=&quot;https:&#x2F;&#x2F;example.com&#x2F;image.jpg&quot; &#x2F;&gt;',
-                    },
-                    {
-                        input: '<script>alert("Hello, world!");</script>',
-                        output: '&lt;script&gt;alert(&quot;Hello, world!&quot;);&lt;&#x2F;script&gt;',
-                    },
-                    {
-                        input: 'Lorem ipsum dolor sit amet',
-                        output: 'Lorem ipsum dolor sit amet',
-                    },
-                    {
-                        input: '',
-                        output: '',
-                    },
-                ];
+                const content = '<p>Hello, world!</p>';
+                const result = preparer.prepare(content, {
+                    ...allOptionsDisabled,
+                    escapeHtml: true,
+                });
 
-                for (const { input, output } of testCases) {
-                    const result = preparer.prepare(input, {
-                        ...allOptionsDisabled,
-                        escapeHtml: true,
-                    });
-                    expect(result).toEqual(output);
-                }
+                expect(result).toEqual('&lt;p&gt;Hello, world!&lt;&#x2F;p&gt;');
             });
 
             it('should not escape HTML by default', () => {
